refactor: uffd: move valid example to a crate
Move the uffd handler example code from tests/host_tools/uffd
to a dedicated `uffd_handler` crate under src/. Only the logic
for the valid uffd is moved, the malicious uffd handler will
be under tests/host_tools/uffd_demo.
