+++ docs/snapshotting/handling-page-faults-on-snapshot-resume.md
@@ -127,7 +127,7 @@ unexpected cases when Firecracker crashes before being able to connect/send data
 
 ### Example
 
-An example of a handler process can be found [here](../../tests/host_tools/uffd/src/bin/valid_handler.rs).
+An implementation of a handler process can be found [here](../../src/uffd_handler).
 The process is designed to tackle faults on a certain address by loading into
 memory the entire region that the address belongs to, but users can choose any
 other behavior that suits their use case best.
