+++ src/uffd_handler/src/lib.rs
@@ -1,0 +1,7 @@
+// Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//! The library crate that defines common helper functions that are generally used in
+//! conjunction with userfaultfd handler.
+
+pub mod common;
